Started implementing sabi_trait attribute.
Implemented attribute parsing.
