+++ abi_stable_derive_lib/src/arenas.rs
@@ -64,6 +64,7 @@ declare_arenas! {
     // metalists: syn::MetaList,
     // visibilities: syn::Visibility,
     // tokenstream: TokenStream,
+    meta_attr: syn::Meta,
     expr: syn::Expr,
     strings: String,
     paths: syn::Path,
