+++ abi_stable_derive_lib/src/sabi_trait.rs
@@ -1,0 +1,1 @@
+mod attribute_parsing;
