+++ abi_stable_derive_lib/src/stable_abi/repr_attrs.rs
